requirements.txt
accelerate==0.31.0
easydict==1.13
evaluate==0.4.1
matplotlib==3.8.4
mmcv==2.2.0
mplcursors==0.5.3
nltk==3.9.1
numpy==2.1.0
pandas==2.2.2
prefixspan==0.5.2
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.2.2
scipy==1.14.1
sentence_transformers==3.0.1
torch==2.3.0
torch_geometric==2.5.3
tqdm==4.66.4
transformers==4.41.2
